Rename WebDAV and Nextcloud providers to WebDavAuth and WebDavPublicLink
